+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - [2026-05-16 22:27] data/data/com.termux/files/home/godmode/load_godmode.py
 - [2026-05-16 22:29] home/godmode/config/prefill.json
+- [2026-05-16 22:33] home/godmode/config/env.sh